# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeMind AI - AI Trading Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /journal:
    get:
      operationId: listJournalEntries
      tags: [journal]
      summary: List trade journal entries
      parameters:
        - in: query
          name: asset
          schema:
            type: string
          required: false
        - in: query
          name: limit
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
  /journal/{id}:
    get:
      operationId: getJournalEntry
      tags: [journal]
      summary: Get a journal entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "404":
          description: Not found
    patch:
      operationId: updateJournalEntry
      tags: [journal]
      summary: Update notes on a journal entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryUpdate"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"

  # ── Portfolio ─────────────────────────────────────────────────
...
  /analysis:
    get:
      operationId: listAnalyses
      tags: [analysis]
      summary: List AI market analyses
      parameters:
        - in: query
          name: asset
          schema:
            type: string
          required: false
        - in: query
          name: limit
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: Market analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketAnalysis"
    post:
      operationId: createAnalysis
      tags: [analysis]
      summary: Request a new AI market analysis for an asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisRequest"
      responses:
        "201":
          description: Analysis created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketAnalysis"
  /analysis/{id}:
    get:
      operationId: getAnalysis
      tags: [analysis]
      summary: Get a specific market analysis
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Market analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketAnalysis"
...
  /coaching/reports:
    get:
      operationId: listCoachingReports
      tags: [coaching]
      summary: List coaching reports
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [daily, weekly, monthly]
          required: false
      responses:
        "200":
          description: Coaching reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoachingReport"
    post:
      operationId: generateCoachingReport
      tags: [coaching]
      summary: Generate a new coaching report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachingReportRequest"
      responses:
        "201":
          description: Report generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingReport"
  /coaching/reports/{id}:
    get:
      operationId: getCoachingReport
      tags: [coaching]
      summary: Get a coaching report
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Coaching report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingReport"

  # ── Trade Replay ───────────────────────────────────────────────
...
  /chat/{sessionId}/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get messages for a chat session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message and get AI response
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"

  # ── Risk Management ────────────────────────────────────────────
...
components:
  schemas:
...
    JournalEntry:
      type: object
      required: [id, tradeId, asset, direction, entryPrice, exitPrice, entryReason, marketConditions, aiRecommendation, userAction, result, pnl, pnlPercent, lessonsLearned, createdAt]
      properties:
        id:
          type: integer
        tradeId:
          type: integer
        asset:
          type: string
        direction:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        exitPrice:
          type: number
        entryReason:
          type: string
        exitReason:
          type: ["string", "null"]
        marketConditions:
          type: string
        aiRecommendation:
          type: string
          enum: [buy, sell, wait, hold]
        userAction:
          type: string
          enum: [followed, ignored, partial]
        result:
          type: string
          enum: [win, loss, breakeven]
        pnl:
          type: number
        pnlPercent:
          type: number
        lessonsLearned:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        behaviorFlags:
          type: array
          items:
            type: string
        createdAt:
          type: string
    JournalEntryUpdate:
      type: object
      properties:
        notes:
          type: string
        lessonsLearned:
          type: string

    # ── Portfolio ────────────────────────────────────────────────
...
    MarketAnalysis:
      type: object
      required: [id, asset, bullishScore, bearishScore, confidenceScore, riskScore, recommendation, summary, evidence, riskAnalysis, macroSignals, sentimentSignals, technicalSignals, onchainSignals, newsSignals, createdAt]
      properties:
        id:
          type: integer
        asset:
          type: string
        bullishScore:
          type: number
        bearishScore:
          type: number
        confidenceScore:
          type: number
        riskScore:
          type: number
        recommendation:
          type: string
          enum: [buy, sell, wait, hold]
        summary:
          type: string
        evidence:
          type: array
          items:
            type: string
        riskAnalysis:
          type: string
        macroSignals:
          type: object
          properties:
            fedPolicy:
              type: string
            macroTrend:
              type: string
            btcCorrelation:
              type: number
            riskMode:
              type: string
              enum: [risk-on, risk-off, neutral]
        sentimentSignals:
          type: object
          properties:
            fearGreedIndex:
              type: number
            longShortRatio:
              type: number
            fundingRate:
              type: number
            crowdPositioning:
              type: string
        technicalSignals:
          type: object
          properties:
            rsi:
              type: number
            macdSignal:
              type: string
              enum: [bullish, bearish, neutral]
            trend:
              type: string
              enum: [uptrend, downtrend, sideways]
            support:
              type: number
            resistance:
              type: number
        onchainSignals:
          type: object
          properties:
            whaleActivity:
              type: string
            etfFlows:
              type: string
            defiTvl:
              type: string
        newsSignals:
          type: object
          properties:
            sentiment:
              type: string
              enum: [positive, negative, neutral]
            keyEvents:
              type: array
              items:
                type: string
        createdAt:
          type: string
    AnalysisRequest:
      type: object
      required: [asset]
      properties:
        asset:
          type: string
...
    CoachingReport:
      type: object
      required: [id, type, title, summary, keyInsights, topMistakes, improvements, recommendations, tradingScore, periodStart, periodEnd, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [daily, weekly, monthly]
        title:
          type: string
        summary:
          type: string
        keyInsights:
          type: array
          items:
            type: string
        topMistakes:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        tradingScore:
          type: number
        winRate:
          type: number
        totalTrades:
          type: integer
        totalPnl:
          type: number
        periodStart:
          type: string
        periodEnd:
          type: string
        createdAt:
          type: string
    CoachingReportRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [daily, weekly, monthly]
        periodStart:
          type: string
        periodEnd:
          type: string

    # ── Trade Replay ─────────────────────────────────────────────
...
    ChatMessage:
      type: object
      required: [id, sessionId, role, content, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    ChatResponse:
      type: object
      required: [userMessage, assistantMessage]
      properties:
        userMessage:
          $ref: "#/components/schemas/ChatMessage"
        assistantMessage:
          $ref: "#/components/schemas/ChatMessage"

    # ── Risk ─────────────────────────────────────────────────────
...