# lib/api-zod/src/generated/types/journalEntryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TradeMind AI - AI Trading Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface JournalEntryUpdate {
  notes?: string;
  lessonsLearned?: string;
}